backend/requirements.txt
# ============================================================
# Jerry The Customer Service Bot Backend — Python Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

# Web Framework
fastapi>=0.109.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
websockets>=12.0,<14.0
python-multipart>=0.0.9,<1.0

# AI / ML
groq>=0.4.2,<1.0
sentence-transformers>=2.7.0,<4.0
scikit-learn>=1.4.0,<2.0
torch>=2.2.0,<3.0
numpy>=1.26.0,<3.0

# Vector Database
pinecone-client>=3.0.3,<5.0

# Database (SQLite for local dev, PostgreSQL for production)
sqlalchemy>=2.0.25,<3.0
aiosqlite>=0.20.0,<1.0
asyncpg>=0.29.0,<1.0

# Caching / Sessions (for future Redis integration)
redis>=5.0.1,<6.0

# Config / Utils
python-dotenv>=1.0.1,<2.0
pydantic>=2.6.0,<3.0
pydantic-settings>=2.2.0,<3.0

# Authentication
PyJWT>=2.8.0,<3.0

# HTTP Client (for Shopify API integration)
httpx>=0.26.0,<1.0

# Shopify (for Phase 2)
ShopifyAPI>=12.6.0,<14.0

# Billing
stripe>=8.0.0,<10.0

# Firewall / File Sanitization
Pillow>=10.0.0,<12.0
filetype>=1.2.0,<2.0
